broke up output tests into package
